# byteutils: remove i from lineReader