A slice already is an index into an array, so we can just mutate data and remove trickiness around tracking i. This way the invariants simplify to always processing from the front of data.

Test Plan: go test