@efritz 2023-03-25
Won't this increase memory again by creating additional slice headers for each loop?

That defeats the purpose of the change in the first place.