@eseliger 2023-03-25
This doesn't add any new allocations, but writing the new slice headers seems slower on larger files vs keeping the index around. I'm not opposed to this change perse if it makes code more readable, and microoptimizations are the evil of all bad, but wanted to put that out here. Let me know what you think, happy either way. 

New

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/internal/byteutils
BenchmarkNewLineReader-10               12561969                90.08 ns/op            0 B/op          0 allocs/op
BenchmarkBytesSplit-10                   9507043               124.3 ns/op           176 B/op          1 allocs/op
BenchmarkNewLineReaderLongLine-10         686592              1664 ns/op               0 B/op          0 allocs/op
BenchmarkBytesSplitLongLine-10            519225              2199 ns/op              24 B/op          1 allocs/op
BenchmarkNewLineReaderManyLines-10          7966            146601 ns/op               0 B/op          0 allocs/op
BenchmarkBytesSplitManyLines-10             7452            149668 ns/op          245760 B/op          1 allocs/op
PASS
```

Old

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/internal/byteutils
BenchmarkNewLineReader-10               13509804                89.83 ns/op            0 B/op          0 allocs/op
BenchmarkBytesSplit-10                   9508216               124.6 ns/op           176 B/op          1 allocs/op
BenchmarkNewLineReaderLongLine-10         699908              1650 ns/op               0 B/op          0 allocs/op
BenchmarkBytesSplitLongLine-10            534580              2189 ns/op              24 B/op          1 allocs/op
BenchmarkNewLineReaderManyLines-10          8350            137206 ns/op               0 B/op          0 allocs/op
BenchmarkBytesSplitManyLines-10             7432            148418 ns/op          245760 B/op          1 allocs/op
PASS
```