@keegancsmith 2023-03-25
The difference was more dramatic for me, 18% slower for NewLineReader. Honestly, I wasn't expecting that. Abandon ship.

```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/internal/byteutils
                          │   old.txt   │               new.txt               │
                          │   sec/op    │   sec/op     vs base                │
NewLineReader-12            51.03n ± 0%   60.19n ± 0%  +17.94% (p=0.000 n=10)
NewLineReaderLongLine-12    1.509µ ± 0%   1.504µ ± 0%   -0.30% (p=0.001 n=10)
NewLineReaderManyLines-12   125.1µ ± 0%   135.3µ ± 0%   +8.14% (p=0.000 n=10)
geomean                     2.128µ        2.305µ        +8.34%

                          │   old.txt    │               new.txt               │
                          │     B/op     │    B/op     vs base                 │
NewLineReader-12            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
NewLineReaderLongLine-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
NewLineReaderManyLines-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                          │   old.txt    │               new.txt               │
                          │  allocs/op   │ allocs/op   vs base                 │
NewLineReader-12            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
NewLineReaderLongLine-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
NewLineReaderManyLines-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```